Add enzyme/webpack entrypoint - export getPluginsForInstalledReact function - update our karma config to use it - update documentation guides to use it
